% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{idrafsd}
\alias{idrafsd}
\title{Fit IDR to training data}
\usage{
idr(y, X, eps, pars = osqpSettings(verbose = FALSE, eps_abs =
  1e-5, eps_rel = 1e-5, max_iter = 10000L), progress = TRUE)
}
\arguments{
\item{y}{numeric vector (the response variable).}

\item{X}{data frame of numeric or ordered factor variables (the regression
covariates).}

\item{eps}{a parameter in (0,0.5) to specify the AFSD order.}

\item{pars}{parameters for quadratic programming optimization (only relevant
if \code{X} has more than one column), set using
\code{\link[osqp]{osqpSettings}}.}

\item{progress}{display progressbar (\code{TRUE}, \code{FALSE} or \code{1},
\code{0})?}
}
\value{
An object of class \code{"idrfit"} containing the following
components:

\item{\code{X}}{data frame of all distinct covariate combinations used for
the fit (note the different order!).}

\item{\code{y}}{list of all observed responses in the training data for
given covariate combinations in \code{X}.}

\item{\code{cdf}}{matrix containing the estimated CDFs, one CDF per row,
evaluated at \code{thresholds} (see next point). The CDF in the \code{i}th
row corredponds to the estimated conditional distribution of the response
given the covariates values in \code{X[i,]}.}

\item{\code{thresholds}}{the thresholds at which the CDFs in \code{cdf} are
evaluated. The entries in \code{cdf[,j]} are the conditional CDFs evaluated
at \code{thresholds[j]}.}

\item{\code{diagnostic}}{list giving a bound on the precision of the CDF
estimation (the maximal downwards-step in the CDF that has been detected)
and the fraction of CDF estimations that were stopped at the iteration
limit \code{max_iter}. Decrease the parameters \code{eps_abs} and/or
\code{eps_rel} or increase \code{max_iter} in \code{pars} to improve the
precision. See \code{\link[osqp]{osqpSettings}} for more optimization
parameters.}

\item{\code{indices}}{ the row indices of the covariates in \code{X} in the
original training dataset (used for in-sample predictions with
\code{\link{predict.idrfit}}).}

\item{\code{constraints}}{ (in multivariate IDR, \code{NULL} otherwise)
matrices giving the order constraints for optimization. Used in
\code{\link{predict.idrfit}}.}
}
\description{
Fits isotonic distributional regression (IDR) to a training
dataset with respect to the AFSD order.
}
\details{
This function computes the isotonic distributional regression (IDR)
of a response \emph{y} on covariates that consists of ensemble forecasts and
are restored in a data-frame \emph{X}. IDR estimates
the cumulative distribution function (CDF) of \emph{y} conditional on
\emph{X} by monotone regression, assuming that \emph{y} is more likely to
take higher values, as \emph{X} increases. Formally, IDR assumes that the
conditional CDF \eqn{F_{y | X = x}(z)} at each fixed threshold \emph{z}
decreases, as \emph{x} increases, or equivalently, that the exceedance
probabilities for any threshold \code{z} \eqn{P(y > z | X = x)} increase
with \emph{x}.

The conditional CDFs are estimated at each threshold in \code{unique(y)}.
This is the set where the CDFs may have jumps. If \code{X} contains more
than one variable, the CDFs are estimated by calling
\code{\link[osqp]{solve_osqp}} from the package \pkg{osqp}
\code{length(unique(y))} times. This might take a while if the training
dataset is large.
}
\note{
The function \code{idr} is only intended for fitting IDR model for a
training dataset and storing the results for further processing, but not
for prediction or evaluation, which is done using the output of
\code{\link{predict.idrfit}}.
}
\seealso{
The S3 method \code{\link{predict.idrfit}} for predictions based on
an IDR fit.
}
