% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{idrsd}
\alias{idrsd}
\title{Fit IDRsd to training data}
\usage{
idrsd(y, X, pars = osqpSettings(verbose = FALSE, eps_abs =
  1e-5, eps_rel = 1e-5, max_iter = 10000L), progress = TRUE)
}
\arguments{
\item{y}{numeric vector (the response variable).}

\item{X}{object that fits to specification in \code{type}. For \code{type = 'idr'}, \code{X} must be an IDR object,
for \code{type = 'ensemble'}, \code{X} must be a data.frame where columns correspond to ensemble members, for \code{type = 'ecdf'}, \code{X} must be a \code{(n x m)} matrix
with \code{n = length(y)} and \code{m = length(grid)}, where each row corresponds to ECDF values evaluated at \code{grid}.
For \code{type = 'dis'} \code{X} is empty, for \code{type = 'normal'} and \code{type = 'normal_ab'} \code{X} is a matrix with 2 columns
which represent mu and sigma parameters of the normal distribution.}

\item{grid}{if \code{type == 'ecdf'}, than \code{grid} is vector of threshold values corresponding to ECDF-values in \code{X}}

\item{dis_func}{if \code{type == 'dis'}, then a cumulative distribution function must be specified along with its distributional parameters.}

\item{type}{default is \code{'ensemble'}. Other possibilities are \code{'idr'}, \code{'ecdf'}, \code{'dis'}, \code{'normal'}, \code{'normal_ab'}}

\item{pars}{parameters for quadratic programming optimization (only relevant
if \code{X} has more than one column), set using
\code{\link[osqp]{osqpSettings}}.}

\item{progress}{display progressbar (\code{TRUE}, \code{FALSE} or \code{1},
\code{0})?}
}
\value{
An object of class \code{"idrcal"} containing the following
components:

\item{\code{X}}{data frame of all distinct values (note the different order!).}

\item{\code{y}}{list of all observed responses in the training data for
given \code{X}.}

\item{\code{cdf}}{matrix containing the estimated CDFs, one CDF per row,
evaluated at \code{thresholds} (see next point). The CDF in the \code{i}th
row corredponds to the estimated conditional distribution of the response
given \code{X[i,]}.}

\item{\code{thresholds}}{the thresholds at which the CDFs in \code{cdf} are
evaluated. The entries in \code{cdf[,j]} are the conditional CDFs evaluated
at \code{thresholds[j]}.}

\item{\code{type}}{type of distributional input data as specified by user}

\item{\code{diagnostic}}{list giving a bound on the precision of the CDF
estimation (the maximal downwards-step in the CDF that has been detected)
and the fraction of CDF estimations that were stopped at the iteration
limit \code{max_iter}. Decrease the parameters \code{eps_abs} and/or
\code{eps_rel} or increase \code{max_iter} in \code{pars} to improve the
precision. See \code{\link[osqp]{osqpSettings}} for more optimization
parameters.}

\item{\code{indices}}{ the row indices of  \code{X} in the
original training dataset (used for in-sample predictions with
\code{\link{predict.idrfit}}).}

\item{\code{constraints}}{ (in multivariate IDR, \code{NULL} otherwise)
matrices giving the order constraints for optimization. Used in
\code{\link{predict.idrfit}}.}
}
\description{
Fits isotonic distributional regression (IDR) to a training
dataset with respect to the stochastic order (SD) order.
}
\details{
This function computes the isotonic distributional regression (IDR)
of a response \code{y} on distributional data \code{X} using stochastic order.
IDRsd estimates the cumulative distribution function (CDF) of \code{y} conditional on
\code{X} under the assumption of isotonicity. The odering on \code{X}
is computed using the stochastic order.
The conditional CDFs are estimated at each threshold in \code{unique(y)}.
This is the set where the CDFs may have jumps.
}
\note{
The function \code{idrsd} is only intended for fitting IDR model for a
training dataset and storing the results for further processing, but not
for prediction or evaluation, which is done using the output of
\code{\link{predict.idrcal}}.
}
\seealso{
The S3 method \code{\link{predict.idrcal}} for predictions based on
an IDR fit.
}
