% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{isodeco_crps}
\alias{isodeco_crps}
\title{CRPS decomposition}
\usage{
isodeco_crps(
  y,
  X = NULL,
  grid = NULL,
  dis_func = NULL,
  type = "ensemble",
  inta = NULL,
  intb = NULL,
  pars = osqpSettings(verbose = FALSE, eps_abs = 1e-05, eps_rel = 1e-05, max_iter =
    10000L),
  progress = TRUE,
  ...
)
}
\arguments{
\item{y}{numeric vector (the response variable).}

\item{X}{depends on \code{type}. For \code{type} = "ensemble", it is a numeric data frame with column number corresponding to ensemble size.
For \code{type = "idr"}, it is an IDR object from isodistrreg-package. For \code{type = "ecdf"}, X is a matrix or a list of ecdf values. For \code{type} = "dis", X is NULL. For \code{type} = "normal", X is NULL.}

\item{grid}{if \code{type = "ecdf"}, than grid is a vector or a list of threshold values corresponding to ECDF-values in  \code{X}.}

\item{dis_func}{if \code{type = "dis"}, then a cumulative distribution function must be specified and distributional parameters must be defined.}

\item{type}{default is 'ensemble'. Other possibilities are 'idr', 'ecdf', 'dis' and 'normal}
}
\value{
A list of CRPS decomposition components: miscalibration (MSC), discrimination (DSC), uncertainty (UNC) and the original CRPS.
}
\description{
Computes the individual components of the iso-based CRPS Decomposition: MSC, DSC and UNC
}
\details{
This function computed the CRPS decomposition of a response vector \emph{y} and a distributional forecast \emph{X}, which can be an ensemble and empirical cumulative distributional function,
a normal distribution or any other other closed form CDF specified by its parameters.
}
\examples{
Simulate data and apply crps_deco to all possible input types

}
