% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isodisregSD-package}
\alias{isodisregSD-package}
\title{Isotonic distributional Regression (IDR) for distributional data}
\description{
Isotonic distributional Regression (IDR) for distributional data is a nonparametric method to
estimate conditional distributions under stochastic order constraints.
}
\section{How does it work?}{

Link to Preprint on ArXiv as soon as available
}

