% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{pit}
\alias{pit}
\alias{pit.idrsd}
\alias{pit.data.frame}
\title{Probability integral transform (PIT)}
\usage{
pit(predictions, y, randomize = TRUE, seed = NULL)

\method{pit}{idrsd}(predictions, y, randomize = TRUE, seed = NULL)

\method{pit}{data.frame}(predictions, y, randomize = TRUE, seed = NULL)
}
\arguments{
\item{predictions}{either an object of class \code{idrsd} (output of
\code{\link{predict.idrcal}}), or a \code{data.frame} of numeric variables. In
the latter case, the PIT is computed using the empirical distribution of
the variables in \code{predictions}.}

\item{y}{a numeric vector of obervations of the same length as the number of
predictions.}

\item{randomize}{PIT values should be randomized at discontinuity points of the
predictive CDF (e.g. at zero for precipitation forecasts). Set \code{
  randomize = TRUE} to randomize.}

\item{seed}{argument to \code{set.seed} for random number generation (if
\code{randomize} is \code{TRUE}).}
}
\value{
Vector of PIT values.
}
\description{
Computes the probability integral transform (PIT) of IDRsd or raw
forecasts.
}
\examples{
data("rain")
require("graphics")


ensemble <- rain[1:(3 * 365), 3:54, drop = FALSE]
y <- rain[1:(3 * 365), "obs"]

fit <- idrsd(y = y, X = ensemble, type = 'ensemble')

## Assess forecast using data of next 2
## years and compare to calibration of the raw ensemble

ytest = rain[(3 * 365 + 1):(5 * 365), "obs"]
ensemble_test = rain[(3 * 365 + 1):(5 * 365), 3:54,drop = FALSE]
predictions <- predict(fit, data = ensemble_test)
idrPit <- pit(predictions, ytest, seed = 123)

rawData <- rain[(3 * 365 + 1):(5 * 365), c("HRES", "CTR", paste0("P", 1:50))]
rawPit <- pit(rawData, ytest, seed = 123)

hist(idrPit, xlab = "Probability Integral Transform",
  ylab = "Density", freq = FALSE, main = "Calibrated ensemble")
hist(rawPit, xlab = "Probability Integral Transform",
  ylab = "Density", freq = FALSE, main = "Raw ensemble")
}
\references{
Gneiting, T., Balabdaoui, F. and Raftery, A. E. (2007), 'Probabilistic
forecasts, calibration and sharpness', Journal of the Royal Statistical
Society: Series B (Statistical Methodology) 69(2), 243-268.
}
\seealso{
\code{\link{predict.idrcal}}
}
