% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot.idrsd}
\alias{plot.idrsd}
\title{Plot IDRsd predictions}
\usage{
\method{plot}{idrsd}(
  x,
  index = 1,
  bounds = TRUE,
  col.cdf = "black",
  col.bounds = "blue",
  lty.cdf = 1,
  lty.bounds = 3,
  xlab = "Threshold",
  ylab = "CDF",
  main = "IDR predictive CDF",
  ...
)
}
\arguments{
\item{x}{object of class \code{idrsd} (output of
\code{\link{predict.idrcal}}).}

\item{index}{index of the prediction in \code{x} for which a plot is desired.}

\item{bounds}{whether the bounds should be plotted or not (see
\code{\link{predict.idrfit}} for details about the meaning of the bounds).}

\item{col.cdf}{color of the predictive CDF.}

\item{col.bounds}{color of the bounds.}

\item{lty.cdf}{linetype of the predictive CDF.}

\item{lty.bounds}{linetype of the CDF bounds.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{main}{main title.}

\item{...}{further arguments to \code{\link{plot.stepfun}} or
\code{\link{plot}}.}
}
\value{
The data based on which the plot is drawn (returned invisible).
}
\description{
Plot an IDRsd predictive CDF.
}
\examples{
data("rain")
require("graphics")


ensemble <- rain[1:(3 * 365), 3:54, drop = FALSE]
y <- rain[1:(3 * 365), "obs"]

## Fit IDRsd and plot the predictive CDF

fit <- idrsd(y = y, X = ensemble, type = 'ensemble')
ensemble_test = rain[(3 * 365 + 1), 3:54,drop = FALSE]
pred <- predict(fit, data = ensemble_test)
plot(pred)
}
\seealso{
\code{\link{predict.idrcal}}
}
