% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predict.idrcal}
\alias{predict.idrcal}
\title{Predict method for IDRsd fits}
\usage{
\method{predict}{idrcal}(
  object,
  data = NULL,
  grid = NULL,
  digits = 3,
  interpolation = "linear",
  ...
)
}
\arguments{
\item{object}{IDRsd fit (object of class \code{"idrcal"}).}

\item{data}{optional object with which to
predict. In-sample predictions are returned if this is omitted.}

\item{digits}{number of decimal places for the predictive CDF.}

\item{...}{included for generic function consistency.}
}
\value{
A list of predictions. Each prediction is a \code{data.frame}
containing the following variables:

\item{\code{points}}{the points where the predictive CDF has jumps.}

\item{\code{cdf}}{the estimated CDF evaluated at the \code{points}.}

\item{\code{lower}, \code{upper}}{ (only for out-of-sample predictions)
bounds for the estimated CDF, see 'Details' above.}

The output has the attribute \code{incomparables}, which gives the indices
of all predictions for which the climatological forecast is returned because
the forecast variables are not comparable to the training data.
}
\description{
Prediction based on IDRsd model fit.
}
\details{
If the variables \code{data} for which predictions are
desired are already contained in the training dataset \code{X} for the fit,
\code{predict.idrfit} returns the corresponding in-sample prediction.
Otherwise monotonicity is used to derive upper and lower bounds for the
predictive CDF, and the predictive CDF is a pointwise average of these
bounds.
}
\seealso{
\code{\link{idrsd}} to fit IDRsd to training data.

\code{\link{cdf}}, \code{\link{qpred}} to evaluate the CDF or quantile
function of IDR predictions.

\code{\link{bscore}}, \code{\link{qscore}}, \code{\link{crps}},
\code{\link{pit}} to compute Brier scores, quantile scores, the CRPS and the
PIT of IDR predictions.

\code{\link[isodistrreg:plot.idrsd]{plot}} to plot IDR predictive CDFs.
}
