% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{qpred}
\alias{qpred}
\alias{qpred.idrsd}
\alias{qpred.data.frame}
\title{Quantile function of IDRsd or raw forecasts}
\usage{
qpred(predictions, quantiles)

\method{qpred}{idrsd}(predictions, quantiles)

\method{qpred}{data.frame}(predictions, quantiles)
}
\arguments{
\item{predictions}{either an object of class \code{idrsd} (output of
\code{\link{predict.idrcal}}), or a \code{data.frame} of numeric variables. In
the latter case, quantiles are computed using the empirical distribution of
the variables in \code{predictions}.}

\item{quantiles}{numeric vector of desired quantiles.}
}
\value{
A matrix of forecasts for the desired quantiles, one column per quantile.
}
\description{
Evaluate the the quantile function of IDRsd predictions or of unprocessed
forecasts in a \code{data.frame}.
}
\details{
The quantiles are defined as lower quantiles, that is,
\deqn{
  q(u) = inf(x: cdf(x) >= u).
}
}
\examples{
## Data from IDR package:
library(isodistrreg)
data("rain")

## IDRsd based on ensemble forecast

ensemble <- rain[1:(3 * 365), 3:54, drop = FALSE]
y <- rain[1:(3 * 365), "obs"]

fit <- idrsd(y = y, X = ensemble, type = 'ensemble')
predictions <- predict(fit)

## Compute 95\%-quantile forecast
qpred(predictions, quantiles = 0.95)

}
\seealso{
\code{\link{predict.idrcal}}, \code{\link{cdf}}, \code{\link{qscore}}
}
