% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{qscore}
\alias{qscore}
\title{Quantile scores for IDRsd or raw forecasts}
\usage{
qscore(predictions, quantiles, y)
}
\arguments{
\item{predictions}{either an object of class \code{idrsd} (output of
\code{\link{predict.idrcal}}), or a \code{data.frame} of numeric variables. In
the latter case, quantiles are computed using the empirical distribution of
the variables in \code{predictions}.}

\item{quantiles}{numeric vector of desired quantiles.}

\item{y}{a numeric vector of obervations of the same length as the number of
predictions, or of length 1. In the latter case, \code{y} will be used
for all predictions.}
}
\value{
A matrix of the quantile scores for the desired quantiles, one column per
quantile.
}
\description{
Computes quantile scores of IDRsd quantile predictions or of quantile
predictions from raw forecasts in a \code{data.frame}.
}
\details{
The quantile score of a forecast \emph{x} for the \emph{u}-quantile is
defined as
\deqn{
2(1{x > y} - u)(x - y),
}
where \emph{y} is the observation. For \emph{u = 1/2}, this equals the mean
absolute error of the median forecast.
}
\examples{
data("rain")

## IDRsd based on ensemble forecast

ensemble <- rain[1:(3 * 365), 3:54, drop = FALSE]
y <- rain[1:(3 * 365), "obs"]

fit <- idrsd(y = y, X = ensemble, type = 'ensemble')

## Compute mean absolute error of the median postprocessed forecast using
## data of the next year (out-of-sample predictions) and compare to raw
## forecast

ytest = rain[(3 * 365 + 1):(4 * 365), "obs"]
ensemble_test = rain[(3 * 365 + 1):(4 * 365), 3:54,drop = FALSE]
predictions <- predict(fit, data = ensemble_test)

idrMAE <- mean(qscore(predictions, 0.5, ytest))
rawMAE <- mean(qscore(ensemble_test, 0.5, ytest))

c("idrsd" = idrMAE, "raw" = rawMAE)

}
\references{
Gneiting, T. and Raftery, A. E. (2007), 'Strictly proper scoring rules,
prediction, and estimation', Journal of the American Statistical Association
102(477), 359-378
}
\seealso{
\code{\link{predict.idrcal}}, \code{\link{qpred}}
}
