% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{idrsd}
\alias{idrsd}
\title{Fit IDR to training data}
\usage{
idr(y, X, pars = osqpSettings(verbose = FALSE, eps_abs =
  1e-5, eps_rel = 1e-5, max_iter = 10000L), progress = TRUE)
}
\arguments{
\item{y}{numeric vector (the response variable).}

\item{X}{object that fits to specification in \emph{type}. For \emph{type = 'idr'}, X must be an IDR object,
for \emph{type = 'ensemble'}, X must be a data.frame, for \emph{type = 'ecdf'}, X must be a matrix,
for \emph{type = 'dis'} X is empty, for \emph{type = 'normal'} and \emph{type = 'normal_ab'} X is a matrix with 2 columns
which represent mu and sigma parameters of the normal distribution}

\item{grid}{if type == 'ecdf', than grid is vector of threshold values corresponding to ECDF-values in X}

\item{dis_func}{if type == 'dis', then a cumulative distribution function must be specified along with its distributional parameters.}

\item{type}{default is 'ensemble'. Other possibilities are 'idr', 'ecdf', 'dis', 'normal', normal_ab}

\item{pars}{parameters for quadratic programming optimization (only relevant
if \code{X} has more than one column), set using
\code{\link[osqp]{osqpSettings}}.}

\item{progress}{display progressbar (\code{TRUE}, \code{FALSE} or \code{1},
\code{0})?}
}
\value{
An object of class \code{"idrcal"} containing the following
components:

\item{\code{X}}{data frame of all distinct values (note the different order!).}

\item{\code{y}}{list of all observed responses in the training data for
given \code{X}.}

\item{\code{cdf}}{matrix containing the estimated CDFs, one CDF per row,
evaluated at \code{thresholds} (see next point). The CDF in the \code{i}th
row corredponds to the estimated conditional distribution of the response
given \code{X[i,]}.}

\item{\code{thresholds}}{the thresholds at which the CDFs in \code{cdf} are
evaluated. The entries in \code{cdf[,j]} are the conditional CDFs evaluated
at \code{thresholds[j]}.}

\item{\code{type}}{type of distributional input data as specified by user}

\item{\code{diagnostic}}{list giving a bound on the precision of the CDF
estimation (the maximal downwards-step in the CDF that has been detected)
and the fraction of CDF estimations that were stopped at the iteration
limit \code{max_iter}. Decrease the parameters \code{eps_abs} and/or
\code{eps_rel} or increase \code{max_iter} in \code{pars} to improve the
precision. See \code{\link[osqp]{osqpSettings}} for more optimization
parameters.}

\item{\code{indices}}{ the row indices of  \code{X} in the
original training dataset (used for in-sample predictions with
\code{\link{predict.idrfit}}).}

\item{\code{constraints}}{ (in multivariate IDR, \code{NULL} otherwise)
matrices giving the order constraints for optimization. Used in
\code{\link{predict.idrfit}}.}
}
\description{
Fits isotonic distributional regression (IDR) to a training
dataset with respect to the stochastic order (SD) order.
}
\details{
This function computes the isotonic distributional regression (IDR)
of a response \emph{y} on distributional data \emph{X}. IDR estimates
the cumulative distribution function (CDF) of \emph{y} conditional on
\emph{X} under the assumption of isotonicity. The odering on \emph{X}
is computed using the stochastic order.
The conditional CDFs are estimated at each threshold in \code{unique(y)}.
This is the set where the CDFs may have jumps.
}
\note{
The function \code{idrsd} is only intended for fitting IDR model for a
training dataset and storing the results for further processing, but not
for prediction or evaluation, which is done using the output of
\code{\link{predict.idrcal}}.
}
\seealso{
The S3 method \code{\link{predict.idrcal}} for predictions based on
an IDR fit.
}
