% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{isodeco_crps}
\alias{isodeco_crps}
\title{CRPS decomposition}
\usage{
isodeco_crps(
  y,
  X = NULL,
  grid = NULL,
  dis_func = NULL,
  type = "ensemble",
  inta = NULL,
  intb = NULL,
  pars = osqpSettings(verbose = FALSE, eps_abs = 1e-05, eps_rel = 1e-05, max_iter =
    10000L),
  progress = TRUE,
  ...
)
}
\arguments{
\item{y}{numeric vector (the response variable).}

\item{X}{object that fits to specification in \code{type}. For \code{type = 'idr'}, \code{X} must be an IDR object,
for \code{type = 'ensemble'}, \code{X} must be a data.frame where columns correspond to ensemble members, for \code{type = 'ecdf'}, \code{X} must be a \code{(n x m)} matrix
with \code{n = length(y)} and \code{m = length(grid)}, where each row corresponds to ECDF values evaluated at \code{grid}.
For \code{type = 'dis'} \code{X} is empty, for \code{type = 'normal'} and \code{type = 'normal_ab'} \code{X} is a matrix with 2 columns
which represent mu and sigma parameters of the normal distribution.}

\item{grid}{if \code{type == 'ecdf'}, than \code{grid} is vector of threshold values corresponding to ECDF-values in \code{X}}

\item{dis_func}{if \code{type == 'dis'}, then a cumulative distribution function must be specified along with its distributional parameters.}

\item{type}{default is \code{'ensemble'}. Other possibilities are \code{'idr'}, \code{'ecdf'}, \code{'dis'}, \code{'normal'}, \code{'normal_ab'}}

\item{pars}{parameters for quadratic programming optimization (only relevant
if \code{X} has more than one column), set using
\code{\link[osqp]{osqpSettings}}.}

\item{progress}{display progressbar (\code{TRUE}, \code{FALSE} or \code{1},
\code{0})?}
}
\value{
A list of CRPS decomposition components: miscalibration (MSC), discrimination (DSC), uncertainty (UNC) and the original CRPS.
}
\description{
Computes the individual components of the iso-based CRPS Decomposition: MSC, DSC and UNC
}
\details{
This function computed the CRPS decomposition of a response vector \emph{y} and a distributional forecast \emph{X}, which can be an ensemble and empirical cumulative distributional function,
a normal distribution or any other other closed form CDF specified by its parameters.
}
\examples{

#' data("rain")

## IDRsd based on ensemble forecast

ensemble <- rain[1:(3 * 365), 3:54, drop = FALSE]
y <- rain[1:(3 * 365), "obs"]

## Compute CRPS decomposition components
crps_deco <- isodeco_crps(y = y, X = ensemble, type = 'ensemble')
print(crps_deco)


}
